feat(trusty-mpm): unify session start with protected-path routing, rename sessions->session, bare tm shortcut (closes #1916)
`tm session start` previously bypassed protected-path/managed-clone
segregation entirely: it ran `prepare_session` and a hand-rolled
`POST /sessions` + raw tmux spawn directly on the given/cwd directory,
even when that directory was a live, non-tm-managed GitHub-backed git
checkout (issue #1916). `tm session new` already routed correctly
through `ManagedNew` -> `spawn_managed` -> `spawn_managed_inproject`
(the #1913 protected-worktree fix).

`session start` now detects project/source info from the resolved
directory via `guided::derive_project` (the same detector the bare `tm`
guided default already uses) and, when it IS a recognized GitHub-backed
git repo, builds and dispatches the IDENTICAL `SessionAction::New`
request through `managed_route::run` -- the exact `spawn_managed`/
`spawn_managed_inproject` code path `session new` uses -- rather than
duplicating the routing logic. A directory that is NOT a recognized git
repo has no live source tree to protect (mirrors what `tm connect`
already documents as a supported case), so the original in-place
`prepare_session` + detached-tmux-start behavior is preserved unchanged
for that case.

Also folds in the two follow-up items from the same issue:
- Renames the `sessions` (plural) CLI subcommand group to `session`
  (singular) throughout the codebase, tests, and shipped skill docs --
  no compatibility alias, since `tm` has no external users.
- Confirms/documents that bare `tm` invoked inside a recognized git repo
  already shortcuts into the same protected managed-spawn flow (via the
  existing guided-default picker's "launch new" path, which posts the
  byte-equivalent `repo_url`/`ref`/`task` to the same
  `POST /api/v1/sessions/managed` endpoint) -- locked in with a new wire-
  capture test so the two entry points cannot silently diverge.

Splits `session.rs`'s `start` handling into a new `commands/session/start`
submodule (SLOC cap) and splits the session-manager MVP CLI-parse tests
out of `tests.rs` into `tests_behavior_d_tests.rs`, both required to stay
under the line-cap gate once the rename incidentally fixed a `/*`-shaped
comment substring that had been masking `tests.rs`'s true SLOC count.
