+++ crates/trusty-mpm/src/bin/tm/commands/session.rs
@@ -4,17 +4,19 @@
 //! breakers, pause, resume, catchup, run, output, instructions) form a cohesive
 //! group that benefits from a dedicated file.
 //! What: the `session` dispatcher function and its private helpers
-//! `emit_managed_alias_notice`, `compose_session_instructions`. The managed
-//! verbs (`new`/`ls`/`send`/`answer`/`attach`/`stop`/`resume`/`decommission`)
-//! route through the shared chat-core layer (`commands::managed_route`); the
-//! id-or-name resolution for both managed and project sessions goes through the
-//! one canonical `client::resolve_target`. `catchup` is handled by the
-//! DOC-28 cutover bridge in the `catchup` submodule.
+//! `emit_managed_alias_notice`, `compose_session_instructions`. `start` is
+//! handled by the [`start`] submodule (protected-path routing, #1916). The
+//! managed verbs (`new`/`ls`/`send`/`answer`/`attach`/`stop`/`resume`/
+//! `decommission`) route through the shared chat-core layer
+//! (`commands::managed_route`); the id-or-name resolution for both managed and
+//! project sessions goes through the one canonical `client::resolve_target`.
+//! `catchup` is handled by the DOC-28 cutover bridge in the `catchup` submodule.
 //! Test: `cli_parses_session_*`, `cli_parses_session_catchup`,
 //! `compose_session_instructions_*` in `tests.rs`.
 
 // CUTOVER BRIDGE submodule — remove post-migration (#1762)
 pub(crate) mod catchup;
+mod start;
 
 use serde::Deserialize;
 
@@ -27,11 +29,14 @@ use crate::types::{EventRow, SessionRow};
 ///
 /// Why: a session is a Claude Code instance; operators start, stop, list,
 /// reap, and inspect them per project from the shell.
-/// What: `Start` posts `POST /sessions` with the project path; `Stop` and
-/// `Resume` are managed-aware (#1218) — they route to the managed runtime-stop/
-/// resume endpoints when the id/name resolves to a managed session, falling back
-/// to the project-session path otherwise; `Info` resolves a session by id or
-/// friendly name; `List` and `Clean` scope to the project directory.
+/// What: `Start` routes through [`start::start_session`] (protected-path
+/// segregation for a recognized GitHub-backed git repo, in-place
+/// `prepare_session` for everything else — see that function's doc for the
+/// #1916 rationale); `Stop` and `Resume` are managed-aware (#1218) — they
+/// route to the managed runtime-stop/resume endpoints when the id/name
+/// resolves to a managed session, falling back to the project-session path
+/// otherwise; `Info` resolves a session by id or friendly name; `List` and
+/// `Clean` scope to the project directory.
 /// Test: `cli_parses_session_start`, `cli_parses_session_stop`,
 /// `cli_parses_session_list`, `cli_parses_session_clean`,
 /// `cli_parses_session_info`.
@@ -41,102 +46,7 @@ pub(crate) async fn session(
     action: SessionAction,
 ) -> anyhow::Result<()> {
     match action {
-        SessionAction::Start { dir } => {
-            let path = resolve_dir(dir)?;
-            // Prepare the custom instructions Claude Code reads at startup:
-            // deploy composed agents to `~/.claude/agents/` and merge the
-            // project CLAUDE.md. This shared prep is what makes a plain
-            // `claude` process behave as a trusty-mpm session.
-            let fw = trusty_mpm::core::paths::FrameworkPaths::default();
-            match trusty_mpm::core::session_launch::prepare_session(&fw, &path) {
-                Ok(report) => {
-                    println!(
-                        "Agents: {} deployed, {} skipped, {} unchanged",
-                        report.deploy.deployed.len(),
-                        report.deploy.skipped.len(),
-                        report.deploy.unchanged.len(),
-                    );
-                    if report.instructions.claude_md_created {
-                        println!("  Created CLAUDE.md stub in {}", path.display());
-                    }
-                    println!(
-                        "Instructions: {} agents in delegation authority",
-                        report.instructions.agent_count
-                    );
-                    println!(
-                        "  Merged instructions written to {}",
-                        report.stash.display()
-                    );
-                    // DOC-28 cutover bridge: print catch-up digest as seed context.
-                    // CUTOVER BRIDGE — remove post-migration (#1762)
-                    if let Some(ctx) = report.catchup_context {
-                        println!("\n---\n\n## Recent Activity (catch-up)\n\n{ctx}");
-                    }
-                }
-                Err(err) => eprintln!("warning: session preparation failed: {err}"),
-            }
-
-            #[derive(Deserialize)]
-            struct Body {
-                #[serde(default)]
-                name: String,
-            }
-            let body: Body = client
-                .post(format!("{url}/sessions"))
-                .json(&serde_json::json!({
-                    "project": path,
-                    "project_path": path,
-                }))
-                .send()
-                .await?
-                .error_for_status()?
-                .json()
-                .await?;
-
-            // The daemon only registers session state now — it no longer
-            // spawns the tmux host (that caused session proliferation). The
-            // CLI owns the actual launch: create a detached tmux session in
-            // the project directory and start `claude` in it.
-            let workdir = path.to_string_lossy().to_string();
-            let new_session = std::process::Command::new("tmux")
-                .args(["new-session", "-d", "-s", &body.name, "-c", &workdir])
-                .status();
-            match new_session {
-                Ok(status) if status.success() => {
-                    let send = std::process::Command::new("tmux")
-                        .args([
-                            "send-keys",
-                            "-t",
-                            &body.name,
-                            &format!(
-                                "claude {}",
-                                trusty_mpm::core::model_inject::PERMISSION_MODE_FLAG
-                            ),
-                            "Enter",
-                        ])
-                        .status();
-                    match send {
-                        Ok(s) if s.success() => {
-                            println!("started session {} (tmux + claude)", body.name);
-                        }
-                        Ok(_) | Err(_) => {
-                            eprintln!(
-                                "warning: tmux session {} created but failed to start claude",
-                                body.name
-                            );
-                            println!("started session {}", body.name);
-                        }
-                    }
-                }
-                Ok(_) | Err(_) => {
-                    eprintln!(
-                        "warning: failed to create tmux session {}; run `claude` manually in {}",
-                        body.name, workdir
-                    );
-                    println!("started session {}", body.name);
-                }
-            }
-        }
+        SessionAction::Start { dir } => start::start_session(client, url, dir).await?,
         SessionAction::Stop { id_or_name } => {
             // #1218: `stop` is managed-aware. If the argument resolves to a
             // MANAGED session (by id or friendly name) via the canonical
@@ -187,7 +97,7 @@ pub(crate) async fn session(
             url: tui_url,
             interval_ms,
         } => {
-            // #1392: the coordinator TUI is `tm sessions tui` (formerly the
+            // #1392: the coordinator TUI is `tm session tui` (formerly the
             // top-level `tm coordinator-tui`). It polls the daemon live, so its
             // `run` is async and runs on the tokio runtime like `tui::run`.
             // `resolve_daemon_url` honours an explicit `--url`, then the lock
@@ -237,8 +147,8 @@ pub(crate) async fn session(
                     // Fall back: the id/name may belong to a MANAGED session
                     // (SM sessions live in the managed store, not the project-session
                     // store). Fetch the managed list and search there too before
-                    // giving up. This fixes the gap where `tm sessions info <uuid>`
-                    // printed "not found" for sessions visible in `tm sessions ls`.
+                    // giving up. This fixes the gap where `tm session info <uuid>`
+                    // printed "not found" for sessions visible in `tm session ls`.
                     let managed = info_from_managed_store(client, url, &id_or_name).await;
                     match managed {
                         Some(val) => println!("{}", serde_json::to_string_pretty(&val)?),
@@ -555,10 +465,10 @@ fn matches_session(session: &serde_json::Value, id_or_name: &str) -> bool {
 
 /// Fetch the managed session list and find a session matching `id_or_name`.
 ///
-/// Why: `tm sessions info` queries the project-session store first; managed
+/// Why: `tm session info` queries the project-session store first; managed
 /// sessions live in a separate store and return 404 there. This fallback
 /// prevents the confusing "not found" message for sessions that ARE visible in
-/// `tm sessions ls`. Non-404 HTTP errors are logged as warnings so daemon 5xx
+/// `tm session ls`. Non-404 HTTP errors are logged as warnings so daemon 5xx
 /// responses are not silently swallowed as "not found".
 /// What: GETs `/api/v1/sessions/managed`, searches by id-exact then name-exact
 /// via `matches_session`, returns the first match as a `serde_json::Value`.
@@ -601,7 +511,7 @@ async fn info_from_managed_store(
 /// prompt — the text actually delivered to `claude --append-system-prompt-file`.
 /// The old code returned `output.merged` (the legacy pipeline: INSTRUCTIONS.md +
 /// delegation authority + CLAUDE.md) for display, while stashing `resolve_pm_prompt`
-/// separately. That caused `tm sessions instructions` to print content that differed
+/// separately. That caused `tm session instructions` to print content that differed
 /// from what Claude received, which is exactly the divergence issue #382 describes.
 /// The single source of truth for "what claude receives" is `resolve_pm_prompt`;
 /// the display and the stash must both come from it.
@@ -634,7 +544,7 @@ pub(crate) fn compose_session_instructions(
     // The single source of truth for the live PM prompt is
     // `build_system_prompt_for`, NOT the bare `resolve_pm_prompt`. The launcher
     // applies HR-4 output-style version-fallback injection on top of the resolved
-    // prompt (issue #1409), so `tm sessions instructions` must show — and the
+    // prompt (issue #1409), so `tm session instructions` must show — and the
     // stash must hold — that SAME injected text. Writing the pre-injection
     // `resolve_pm_prompt` here made the display/stash diverge from the real launch
     // prompt whenever `claude` was absent/old (injection fires), the same #382
