+++ crates/trusty-mpm/src/client/http_client/types.rs
@@ -403,7 +403,7 @@ pub struct ManagedSessionSummary {
     pub source_id: Option<String>,
     /// Task description for the session (additive field; absent for legacy records).
     ///
-    /// Why: the `tm sessions ls` table needs a task column to distinguish sessions
+    /// Why: the `tm session ls` table needs a task column to distinguish sessions
     /// without attaching; this mirrors the `SessionRecord::task` field now surfaced
     /// in `SessionSummary`. Old daemon responses without the field deserialize as
     /// `None`.
@@ -411,7 +411,7 @@ pub struct ManagedSessionSummary {
     pub task: Option<String>,
     /// Working directory for the session (additive; absent for legacy records).
     ///
-    /// Why: `tm sessions info` and the ls table need to show where the session is
+    /// Why: `tm session info` and the ls table need to show where the session is
     /// running; this mirrors `SessionRecord::cwd` now surfaced in `SessionSummary`.
     /// Old daemon responses without the field deserialize as `None`.
     #[serde(default)]
